cros: Correct timer_get_us() prototype
This prototype needs to move to common.h, but the in meantime, just
fix it so that it uses ulong as the return value.

BUG=none
BRANCH=none
TEST=manual
build and boot on snow; build for link

Reviewed-on: https://gerrit.chromium.org/gerrit/30852

+++ cros/include/common.h
@@ -55,6 +55,6 @@ void *cros_memalign_cache(size_t n);
 int display_clear(void);
 
 /* put this prototype here for now */
-unsigned timer_get_us(void);
+ulong timer_get_us(void);
 
 #endif /* CHROMEOS_COMMON_H_ */
